requirements.txt
asgiref==3.7.2
distlib==0.3.6
Django==4.2.2
django-crispy-forms==1.14.0
filelock==3.12.0
Pillow==9.5.0
platformdirs==3.5.1
sqlparse==0.4.4
virtualenv==20.23.0
